ui: move shared tokens out of widgets package
